Refactor sponsor components to use PlatinumSponsors instead of Sponsors
