+++ components/sponsors/PlatinumSponsorsList.ts
